add load config
